mv _create_first_user to security module
